refactor: change project name
Since the project is a todo web application it should reflect as is.
