+++ src/main/java/com/training/todo/HelloServlet.java
@@ -1,4 +1,4 @@
-package com.training.tomcatdemosample;
+package com.training.todo;
 
 import java.io.*;
